Move MinHeap.java and MaxHeap.java to the "FixedArrayImplementation" folder within the test folder.
